loan rejection email

+++ Small_Business_Finance_App/src/main/java/com/infy/Small_Business_Finance_App/app/model/Email.java
@@ -10,7 +10,7 @@ public class Email
 	private String subject;
 	private String text;
 	private String borrowerName;
-	private Integer applicationNo;
+	private Long applicationNo;
 	private Double loanAmount;
 	private String tenure;
 	private Float intRate;
